+++ src/composables/data.ts
@@ -59,6 +59,7 @@ const vatsim = {
     mandatoryData: shallowRef<VatsimMandatoryConvertedData | null>(null),
     versions: ref<VatDataVersions['vatsim'] | null>(null),
     updateTimestamp: ref(''),
+    updateTime: ref(0),
 };
 
 export function useDataStore() {
@@ -80,6 +81,8 @@ export function setVatsimDataStore(vatsimData: VatsimLiveDataShort) {
 }
 
 export function setVatsimMandatoryData(data: VatsimMandatoryData) {
+    time.value = data.serverTime;
+    vatsim.updateTime.value = data.timestampNum;
     vatsim.mandatoryData.value = {
         pilots: data.pilots.map(([cid, lon, lat, icon, heading]) => {
             const coords = fromLonLat([lon, lat]);