+++ src/types/data/vatsim.ts
@@ -145,6 +145,8 @@ export type VatsimShortenedData = {
 
 export type VatsimMandatoryData = {
     timestamp: string;
+    timestampNum: number;
+    serverTime: number;
     pilots: [cid: VatsimPilot['cid'], longitude: VatsimPilot['longitude'], latitude: VatsimPilot['latitude'], icon: AircraftIcon, heading: number][];
     controllers: [VatsimController['cid'], VatsimController['callsign'], VatsimController['frequency'], VatsimController['facility']][];
     atis: VatsimMandatoryData['controllers'];