+++ src/utils/backend/worker/data-worker.ts
@@ -286,6 +286,35 @@ CronJob.from({
 
             await updateVatsimExtendedPilots();
 
+            radarStorage.vatsim.data.controllers.push({
+                callsign: 'BIRD_CTR',
+                cid: 3,
+                facility: (await import('~/utils/data/vatsim')).useFacilitiesIds().CTR,
+                frequency: '122.122',
+                last_updated: '',
+                logon_time: '',
+                name: '',
+                rating: 0,
+                server: '',
+                text_atis: ['test3'],
+                visual_range: 0,
+            });
+
+            radarStorage.vatsim.data.controllers.push({
+                callsign: 'BIRD_S_CTR',
+                cid: 4,
+                facility: (await import('~/utils/data/vatsim')).useFacilitiesIds().CTR,
+                frequency: '122.122',
+                last_updated: '',
+                logon_time: '',
+                name: '',
+                rating: 0,
+                server: '',
+                text_atis: ['test3'],
+                visual_range: 0,
+            });
+
+
             /*            radarStorage.vatsim.data.atis.push({
                 callsign: 'MSK_APP',
                 cid: 3,